Take out serialization to its own module
