+++ http_client/src/main/kotlin/Client.kt
@@ -3,7 +3,7 @@ package com.hexagonkt.http.client
 import com.hexagonkt.http.Cookie
 import com.hexagonkt.http.Method.*
 import com.hexagonkt.http.Path
-import com.hexagonkt.core.serialization.SerializationFormat
+import com.hexagonkt.serialization.SerializationFormat
 
 /**
  * Client to use other REST services.
