+++ http_client/src/test/kotlin/ClientTest.kt
@@ -10,7 +10,7 @@ import com.hexagonkt.http.server.Server
 import com.hexagonkt.http.server.ServerSettings
 import com.hexagonkt.http.server.jetty.JettyServletAdapter
 import com.hexagonkt.http.server.serve
-import com.hexagonkt.core.serialization.*
+import com.hexagonkt.serialization.*
 import com.hexagonkt.serialization.json.Json
 import com.hexagonkt.serialization.yaml.Yaml
 import org.junit.jupiter.api.*
