+++ http_client_ahc/src/main/kotlin/AhcAdapter.kt
@@ -3,8 +3,8 @@ package com.hexagonkt.http.client.ahc
 import com.hexagonkt.core.logging.Logger
 import com.hexagonkt.core.helpers.ensureSize
 import com.hexagonkt.core.helpers.fail
-import com.hexagonkt.core.serialization.SerializationManager.formatOf
-import com.hexagonkt.core.serialization.serialize
+import com.hexagonkt.serialization.SerializationManager.formatOf
+import com.hexagonkt.serialization.serialize
 import com.hexagonkt.http.Cookie
 import com.hexagonkt.http.Method
 import com.hexagonkt.http.Method.*
