+++ http_server/src/main/kotlin/Request.kt
@@ -3,11 +3,11 @@ package com.hexagonkt.http.server
 import com.hexagonkt.http.Cookie
 import com.hexagonkt.http.Method
 import com.hexagonkt.http.Part
-import com.hexagonkt.core.serialization.SerializationFormat
-import com.hexagonkt.core.serialization.SerializationManager
-import com.hexagonkt.core.serialization.SerializationManager.requireDefaultFormat
-import com.hexagonkt.core.serialization.parse
-import com.hexagonkt.core.serialization.parseObjects
+import com.hexagonkt.serialization.SerializationFormat
+import com.hexagonkt.serialization.SerializationManager
+import com.hexagonkt.serialization.SerializationManager.requireDefaultFormat
+import com.hexagonkt.serialization.parse
+import com.hexagonkt.serialization.parseObjects
 import java.security.cert.X509Certificate
 import kotlin.reflect.KClass
