+++ http_server/src/test/kotlin/PortHttpServerSamplesTest.kt
@@ -11,7 +11,7 @@ import com.hexagonkt.http.client.ahc.AhcAdapter
 import com.hexagonkt.http.server.ServerFeature.SESSIONS
 import com.hexagonkt.serialization.json.JacksonMapper
 import com.hexagonkt.serialization.json.Json
-import com.hexagonkt.core.serialization.SerializationManager
+import com.hexagonkt.serialization.SerializationManager
 import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.TestInstance
