+++ http_server/src/test/kotlin/RequestTest.kt
@@ -5,8 +5,8 @@ import com.fasterxml.jackson.annotation.JsonProperty
 import com.hexagonkt.http.server.test.TestRequest
 import com.hexagonkt.http.server.test.testCall
 import com.hexagonkt.serialization.json.JacksonMapper
-import com.hexagonkt.core.serialization.SerializationManager
-import com.hexagonkt.core.serialization.toObject
+import com.hexagonkt.serialization.SerializationManager
+import com.hexagonkt.serialization.toObject
 import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.TestInstance
