+++ http_server/src/test/kotlin/examples/GenericTest.kt
@@ -12,9 +12,9 @@ import com.hexagonkt.http.server.Server
 import com.hexagonkt.http.server.ServerPort
 import com.hexagonkt.serialization.json.JacksonMapper
 import com.hexagonkt.serialization.json.Json
-import com.hexagonkt.core.serialization.SerializationManager
-import com.hexagonkt.core.serialization.toObject
-import com.hexagonkt.core.serialization.parse
+import com.hexagonkt.serialization.SerializationManager
+import com.hexagonkt.serialization.toObject
+import com.hexagonkt.serialization.parse
 import org.junit.jupiter.api.AfterAll
 import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.Test
