+++ serialization/src/main/resources/serialization/mime.types
