+++ serialization/src/test/kotlin/SerializationPackageTest.kt
@@ -1,4 +1,4 @@
-package com.hexagonkt.core.serialization
+package com.hexagonkt.serialization
 
 import com.hexagonkt.core.helpers.toStream
 import com.hexagonkt.serialization.json.JacksonMapper
