+++ serialization/src/test/kotlin/TestResourceBundles.kt
@@ -1,5 +1,5 @@
 @file:Suppress("ClassName", "unused")
-package com.hexagonkt.core.serialization
+package com.hexagonkt.serialization
 
 import java.net.URL
