+++ serialization/src/test/resources/sample.yml
