+++ serialization/src/test/resources/sample_es.yml
