+++ serialization_csv/src/main/kotlin/Csv.kt
@@ -7,7 +7,7 @@ import com.fasterxml.jackson.dataformat.csv.CsvMapper
 import com.fasterxml.jackson.dataformat.csv.CsvParser
 import com.hexagonkt.serialization.json.JacksonHelper
 import com.hexagonkt.serialization.json.JacksonHelper.setupObjectMapper
-import com.hexagonkt.core.serialization.SerializationFormat
+import com.hexagonkt.serialization.SerializationFormat
 import java.io.InputStream
 import java.io.OutputStream
 import kotlin.reflect.KClass
