+++ serialization_json/src/main/kotlin/JacksonMapper.kt
@@ -1,6 +1,6 @@
 package com.hexagonkt.serialization.json
 
-import com.hexagonkt.core.serialization.Mapper
+import com.hexagonkt.serialization.Mapper
 import kotlin.reflect.KClass
 
 object JacksonMapper : Mapper {
