+++ serialization_xml/src/test/kotlin/XmlTest.kt
@@ -3,7 +3,7 @@ package com.hexagonkt.serialization.xml
 import com.hexagonkt.core.helpers.get
 import com.hexagonkt.core.helpers.println
 import com.hexagonkt.core.helpers.toStream
-import com.hexagonkt.core.serialization.*
+import com.hexagonkt.serialization.*
 import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.TestInstance
