+++ serialization_yaml/src/test/kotlin/YamlTest.kt
@@ -1,7 +1,7 @@
 package com.hexagonkt.serialization.yaml
 
 import com.hexagonkt.core.helpers.toStream
-import com.hexagonkt.core.serialization.*
+import com.hexagonkt.serialization.*
 import com.hexagonkt.serialization.json.Json
 import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.Test
